requirement.txt
boto3
google-cloud-storage
pytest
coverage
